chore: add validations related to Carrinho crud operations

+++ src/services/NSE.Carrinho.API/Controllers/CarrinhoController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -38,6 +39,7 @@ namespace NSE.Carrinho.API.Controllers
             else
                 ManipularCarrinhoExistente(carrinho, item);
 
+            ValidarCarrinho(carrinho);
             if (!OperacaoValida()) return CustomResponse();
 
             await PersistirDados();
@@ -58,6 +60,9 @@ namespace NSE.Carrinho.API.Controllers
 
             carrinho.AtualizarUnidades(itemCarrinho, item.Quantidade);
 
+            ValidarCarrinho(carrinho);
+            if (!OperacaoValida()) return CustomResponse();
+
             _context.CarrinhoItens.Update(itemCarrinho);
             _context.CarrinhoCliente.Update(carrinho);
 
@@ -74,7 +79,11 @@ namespace NSE.Carrinho.API.Controllers
             var itemCarrinho = await ObterItemCarrinhoValidado(produtoId, carrinho);
             if (itemCarrinho == null) return CustomResponse();
 
+            ValidarCarrinho(carrinho);
+            if (!OperacaoValida()) return CustomResponse();
+
             carrinho.RemoverItem(itemCarrinho);
+
             _context.CarrinhoItens.Remove(itemCarrinho);
             _context.CarrinhoCliente.Update(carrinho);
 
@@ -149,6 +158,17 @@ namespace NSE.Carrinho.API.Controllers
             var result = await _context.SaveChangesAsync();
             if (result <= 0)
                 AdicionarErroProcessamento("Não foi possível persistir os dados no banco");
+        }
+
+        private bool ValidarCarrinho(CarrinhoCliente carrinho)
+        {
+            if (carrinho.EhValido())
+                return true;
+
+            carrinho.ValidationResult.Errors.ToList()
+                .ForEach(e => AdicionarErroProcessamento(e.ErrorMessage));
+
+            return false;
         }
     }
 }
